+++ chaosgcp/cloudrun/actions.py
@@ -105,6 +105,7 @@ def update_service(
     traffic: List[Dict[str, Any]] = None,
     labels: Dict[str, str] = None,
     annotations: Dict[str, str] = None,
+    vpc_access_config: Dict[str, str] = None,
     configuration: Configuration = None,
     secrets: Secrets = None,
 ):
@@ -146,9 +147,12 @@ def update_service(
     :param annotations: optional annotations to set on the service
     :param configuration:
     :param secrets:
+    :param vpc_access_config: optional value for vpc_connect
 
     :return:
-    """  # noqa: E501
+    """
+
+    # noqa: E501
     credentials = load_credentials(secrets)
 
     traffics = None
@@ -159,13 +163,20 @@ def update_service(
     if container:
         containers = [run_v2.Container(**container)]
 
+    vpc_access = None
+    if vpc_access_config:
+        vpc_access = vpc_access_config
+
     client = run_v2.ServicesClient(credentials=credentials)
+
     tpl = run_v2.RevisionTemplate(
         max_instance_request_concurrency=max_instance_request_concurrency,
         service_account=service_account,
         encryption_key=encryption_key,
         containers=containers,
+        vpc_access=vpc_access,
     )
+
     svc = run_v2.Service(
         name=parent,
         labels=labels,
@@ -177,4 +188,5 @@ def update_service(
 
     operation = client.update_service(request=request)
     response = operation.result()
+
     return response.__class__.to_dict(response)